Redirect after movie creation and clean up logs
Updated the movie creation flow to redirect to the home page after a movie is created instead of ending the response. Removed unnecessary console.log statements from controllers and models. Fixed the navigation link for creating a post to point to '/movies/create' and made minor improvements to comments and server startup message.